陈柏林→修改获取总价接口

+++ src/main/java/com/weisen/www/code/yjf/shopmall/service/impl/Rewrite_SpecificationsServiceImpl.java
@@ -102,7 +102,7 @@ public class Rewrite_SpecificationsServiceImpl implements Rewrite_Specifications
         else {
             Optional<Specifications> byId = rewrite_SpecificationsRepository.findById(id);
             Specifications specifications = byId.get();
-            return Result.suc("获取成功",Integer.valueOf(specifications.getPrice()) * number);
+            return Result.suc("获取成功",Double.valueOf(specifications.getPrice()) * number);
         }
     }
